Introduce 'Actions', repeated experiments
This commit introduces 'Actions', a record of what each actor in the
system has done. The notion here is that we'll use the two actor queues
-- called SENDER_Q and RECEIVER_Q -- to debug experiment runs that go
sideways. This commit also incorporates results from the Action queues
which mean we now are much more likely to run to a successful completion.
As a result, the old main loop is more or less split out into an
'experiment' run and the main loop just runs this over and over until we
hit catastrophic failure.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>