Delay while polling for work
In this commit we remove the TOTAL_SENDERS check and instead delay
in the receiver loop when we're unable to find new work, up to
some limit. This means that once work is unavailable -- because the
sender has hung up -- we have a bounded amount of time to wait before
the receiver will hang up as well. The bound here is quite high but
works well enough in practice.

Next trick will be to back off SeqCst.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>