Revamp experimentation setup, add runtime flag options
This commit fiddles a fair bit with the surrounding code to run the
sender / receiver pair -- removing the Action structs etc -- which
gets us closer to a clean-room situation. Senders now write in a
way that excludes other writers to avoid situations where an offset
will be taken, the scheduler bumps that thread, the newly scheduled
thread flushes and informs the receiver before the original thread is
able to write its content, resulting in a zero to the receiver. To
further test for this we no longer allow zero-value to be written.

A feature flag to allow you to switch between built-in mutex and
parking_lot would be nifty.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>